Improve datetime formatting for database search in delete appointment
modifies the `find_one` method in the appointment_collection by using a single `formatted_datetime` variable instead of separate date and time. A datetime string is formatted from the data and then converted into a datetime object; afterward it is again converted into a required string format to ensure optimal compatibility with the existing database structure.